Usar double en vez de BigDecimal en el scaffold
BigDecimal no se vio en el curso (sesiones 05/06 usan double para
dinero); tambien se ajusta el TODO de Cart para sugerir solo
containsKey/get/put del Map, sin getOrDefault/entrySet.

+++ shop-cart-app/src/main/java/org/example/Product.java
@@ -1,7 +1,5 @@
 package org.example;
 
-import java.math.BigDecimal;
-
 /**
  * Base type for anything that can be added to a {@link Cart}.
  *
@@ -19,10 +17,10 @@ public abstract class Product {
      * Shipping cost for buying {@code quantity} units of this product.
      * Implement per subclass — no instanceof/type checks here or at call sites.
      */
-    public abstract BigDecimal shippingCost(int quantity);
+    public abstract double shippingCost(int quantity);
 
     /**
      * Tax rate applied to this product's price (e.g. 0.16 for 16%).
      */
-    public abstract BigDecimal taxRate();
+    public abstract double taxRate();
 }
